Add tests/ dir to reference app
We add a tests/ folder that seems to cover most application layouts [1],
although other variations may exist.

This is ahead of preparing the reference repo for mutation testing, and
it seems easier to have the convention of putting tests in a folder
instead of enforcing test files be prepended with `test_`--although
that would be nice too.

[1] https://realpython.com/python-application-layouts/

(DEVOPS-625)